Reworked structure of xmlrpc library. Now it is dissolved in roscxx
