+++ src/xmlrpc/COPYING
