+++ src/xmlrpc/b64encode.cpp
@@ -1,0 +1,155 @@
+/*
+cencoder.c - c source to a base64 encoding algorithm implementation
+
+This is part of the libb64 project, and has been placed in the public domain.
+For details, see http://sourceforge.net/projects/libb64
+*/
+
+#include "b64/encode.h"
+
+namespace base64 {
+
+constexpr int CHARS_PER_LINE = 72;
+
+void base64_init_encodestate(base64_encodestate* state_in)
+{
+  state_in->step = step_A;
+  state_in->result = 0;
+  state_in->stepcount = 0;
+}
+
+char base64_encode_value(char value_in)
+{
+  static const char* encoding = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
+  if (value_in > 63)
+    return '=';
+  return encoding[(int)value_in];
+}
+
+int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in)
+{
+  const char* plainchar = plaintext_in;
+  const char* const plaintextend = plaintext_in + length_in;
+  char* codechar = code_out;
+  char result;
+  char fragment;
+
+  result = state_in->result;
+
+  switch (state_in->step) {
+    while (1) {
+      case step_A:
+        if (plainchar == plaintextend) {
+          state_in->result = result;
+          state_in->step = step_A;
+          return codechar - code_out;
+        }
+      fragment = *plainchar++;
+      result = (fragment & 0x0fc) >> 2;
+      *codechar++ = base64_encode_value(result);
+      result = (fragment & 0x003) << 4;
+      [[fallthrough]];
+      case step_B:
+        if (plainchar == plaintextend) {
+          state_in->result = result;
+          state_in->step = step_B;
+          return codechar - code_out;
+        }
+      fragment = *plainchar++;
+      result |= (fragment & 0x0f0) >> 4;
+      *codechar++ = base64_encode_value(result);
+      result = (fragment & 0x00f) << 2;
+      [[fallthrough]];
+      case step_C:
+        if (plainchar == plaintextend) {
+          state_in->result = result;
+          state_in->step = step_C;
+          return codechar - code_out;
+        }
+      fragment = *plainchar++;
+      result |= (fragment & 0x0c0) >> 6;
+      *codechar++ = base64_encode_value(result);
+      result = (fragment & 0x03f) >> 0;
+      *codechar++ = base64_encode_value(result);
+
+      ++(state_in->stepcount);
+      if (state_in->stepcount == CHARS_PER_LINE / 4) {
+        *codechar++ = '\n';
+        state_in->stepcount = 0;
+      }
+    }
+  }
+  /* control should not reach here */
+  return codechar - code_out;
+}
+
+int base64_encode_blockend(char* code_out, base64_encodestate* state_in)
+{
+  char* codechar = code_out;
+
+  switch (state_in->step) {
+    case step_B:
+      *codechar++ = base64_encode_value(state_in->result);
+    *codechar++ = '=';
+    *codechar++ = '=';
+    break;
+    case step_C:
+      *codechar++ = base64_encode_value(state_in->result);
+    *codechar++ = '=';
+    break;
+    case step_A:
+      break;
+  }
+  *codechar++ = '\n';
+
+  return codechar - code_out;
+}
+
+Encoder::Encoder(int buffersize_in) : _buffersize(buffersize_in)
+{
+  base64_init_encodestate(&_state);
+}
+
+int Encoder::encode(char value_in)
+{
+  return base64_encode_value(value_in);
+}
+
+int Encoder::encode(const char* code_in, const int length_in, char* plaintext_out)
+{
+  return base64_encode_block(code_in, length_in, plaintext_out, &_state);
+}
+
+int Encoder::encode_end(char* plaintext_out)
+{
+  return base64_encode_blockend(plaintext_out, &_state);
+}
+
+void Encoder::encode(std::istream& istream_in, std::ostream& ostream_in)
+{
+  base64_init_encodestate(&_state);
+  //
+  const int N = _buffersize;
+  char* plaintext = new char[N];
+  char* code = new char[2 * N];
+  int plainlength;
+  int codelength;
+
+  do {
+    istream_in.read(plaintext, N);
+    plainlength = istream_in.gcount();
+    //
+    codelength = encode(plaintext, plainlength, code);
+    ostream_in.write(code, codelength);
+  } while (istream_in.good() && plainlength > 0);
+
+  codelength = encode_end(code);
+  ostream_in.write(code, codelength);
+  //
+  base64_init_encodestate(&_state);
+
+  delete[] code;
+  delete[] plaintext;
+}
+
+} // namespace base64
