--- xmlrpcpp/libb64/INSTALL
@@ -1,44 +1,0 @@
-libb64: Base64 Encoding/Decoding Routines
-======================================
-
-Requirements:
-------------
-This piece of software has minimal requirements.
-
-I have tested it on the following systems:
-
-- a Linux machine, with the following specs:
-(this was the original development machine)
-	* FedoraCore 4
-	* kernel v. 2.6.11 (stock FC4 kernel)
-	* gcc version 4.0.1 20050727 (Red Hat 4.0.1-5)
-	* glibc-2.3.5-10
-	* make v. 3.80
-	* some arb version of makedepend
-
-- Windows XP machine
-	* MSYS 1.0
-	* MinGW 5.1.4
-	* gcc version 3.4.5 (mingw-vista special r3)
-
-- Windows XP machine (same as above)
-	* Microsoft Visual Studio 2010, Version 10.0.30319.1 RTMRel
-	
-Barring any serious screwups on my part, this code should compile and run sweetly
-under Cygwin and other systems too. If you DO get it running under some weird arch/os setup,
-send me a mail, please.
-
-Compiling:
----------
-There is no configure. It would be overkill for something so simple...
-Run make in the root directory.
-
-Installing:
-----------
-Since the current targets are a standalone executable and a static library
-(fancy name for archive) with some headers, an install script has not been implemented yet.
-Simply copy the executable into your path, and use it.
-
---
-peace out
-Chris
